// ProyectoFinGradoFrontend/DiarioEntrenamiento/src/app/shared/Models/PerimetrosRequest.ts
export interface PerimetrosRequest {
  cuello: number | null;
  brazoDchoRelajado: number | null;
  brazoDchoTension: number | null;
  brazoIzqRelajado: number | null;
  brazoIzqTension: number | null;
  pecho: number | null;
  hombro: number | null;
  cintura: number | null;
  cadera: number | null;
  abdomen: number | null;
  musloDcho: number | null;
  musloIzq: number | null;
  pantorrillaDcha: number | null;
  pantorrillaIzq: number | null;
}